requirements.txt
python-dotenv
asyncpg
aiogram
sqlalchemy
gino